refactor: navLinkのimportをconfig/link/linkNavsからconfig/linkに変更した
